refactor tables config types
